MENT-213: changes in my_win_unlink() made rpl_mysqldump_slave to fail
When the binary log file is moved to TEMP dir on Windows (see MENT-216), and
TEMP dir is located not at the same volume as the binary log file, MoveFile()
function returns "success", the file is copied to the destination location,
but it is not removed from the source location, what causes MENT-213.

The fix is to rollback the changes in my_win_unlink(). Initially the changes
were introduced in MENT-9, so this commint can be squashed with MENT-9.