MENT-482: main.backup_log fails with --ps-protocol mtr option
Table opening order is different for mysql_create_like_table() and
mysql_test_create_table()(which is invoked from check_prepared_statement()).

That is why when "CREATE TABLE IF NOT EXISTS t_exists LIKE v" is
executed, 't_exists' table is opened firstly, and, as it exists, the
statement execution returns 'ok' and does not create table.

But when the statement is prepared, 'v' table is opened firstly, and as
it does not exist, the error occured.

The fix is to have existing table in 'LIKE' part of the statement to
avoid error when ps protocol is used for mtr testing.